Silver/Watchers/requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# WhatsApp watcher (Green API)
requests>=2.28.0
